chore: move out warm_up_cache_for_snapshot_constructors and related functionaly to snapshotter_utils for reuse
