+++ pooler/callback_modules/settings/settings.example.json
@@ -46,8 +46,5 @@
         "0x1f9840a85d5af5bf1d1762f925bdaddc4201f984",
         "0x2260fac5e5542a773aa44fbcfedf7c193bc2c599",
         "0x956f47f50a910163d8bf957cf5846d573e7f87ca"
-    ],
-    "uniswap_functions": {
-        "retrial_attempts": 5
-    }
+    ]
 }
